Modified Chain to use TypeConverters.
